Fix email init function
Email init function uses "merge" function to extend default config file.
It turnes out, that this file is not copied on merge, but mutated changing the original, too.
This caused a problem, when many initEmailService functions are called. They all extend the same
config, which results in all of them binding to the same field.