Add a fix to session id generation
The logic for session id generation and logic that attaches a subscriber to session id field were working in two parallel function,
which could lead to a random order of execution. One option being the fields being created after the method attempts to add a subscriber
which makes the field unusable by the js-sdk module and breaks the functionality of the whole system.