Add a filter for external relevance check
Sometimes we have to consider some external conditions when deciding if the results of
an address check are still valid for the form.
E.g. when the form was hidden, because a logically different address was selected (not in a form of an addressform)
This is the case in JTL Shop. There you can select a delivery address template, which will hide the addressform of the
current delivery address, while triggering the addresscheck in the process (blur events are thrown).
Using the newly created filter its possible to recognize this scenarion and prevent the appearance of
irrelevant address correction modal.

Ref: J5P-50