Use value API that subscriber provides instead of direct update
For some reason i've started using updateValue method instead of setting the value
vie subscriber.value API (which does the nessesary conversions). I'm not sure why, but probably
because I had some buggy behaviour.

The problem with this approach is tho, that those conversions that .value ist taking
of are very important. In fact its impossible to update the country selector without them and
js-sdk goes into lock state where its impossible to change the value of country selector no matter
what one does.

The fix returns the original 1.8.1 state of the subscriber API however all the updates, would update
the internal state of the listener (which seemed as a problem back then, but now I dont really understand why)

Updating the internal state means that if a value is set via subscriber.value, then the internal listener would just
update it in the DOM, but not trigger a change event. Which is fine, because only the address object is using the .value API
at the moment.

Ref: JS-65