Add self reference as parameter to resolvers
In some cases it might be relevant for a resolver to have a reference to subscriber, because
this way the resolver could get a reference to actual DOMElement being observed and implement a custom
read or write logic depending on the circumstances.
For example in showpware 6 the select with country states is initially empty, containing only one element
but later its populated with more options. However the original element before this population contains
the id of the option that will be marked as selected in a custom attribute data-something.
So if the reference to DOM can be retrieved in the resolver, that its possible to implement a logic that would read
from this custom attribute, if no options are provided (yet).

Ref: S6A-211