Add template engine and create view for two use cases
In order to recreate specific conditions we need a way to use js-sdk on separate pages. For each use case
there should be a page. On top of that the pages should be able to display submitted data.
In order to satisfy those requirement besides the express.js we also need to use a templating engine, that will allow
use to keep the front end managable by providing us with extendable layout, macros and includable components.
In this commit we install those dependencies, extend the app.js and implement a use case for country mapping tests.