Update dependencies
In order to ensure that we can work with modern version of node and npm some of the dependencies need an update. Also the SCSS files had to be updated to support newer version of SCSS.

Tested with Node v20, npm 10.8

After updating browser sync there is an issue with a critical vulnerability in one of its dependencies. I estimate it not critical for us, as browser-sync is used mainly for development purposes and is not part of the build buindle. Automatic fix downgrade to an older version, which has 32 vulnerabilities in its dependencies, so we'll just let it be for now.