Add config option to deactivate address autocomplete
For some customers and in some countries the implementation of field based autocomplete (and resulting browser autocomplete deactivation) is not
wished. Essentially they are not happy with the quality or the autocomplete should work differently in that country in comparison to germany.

So as a temporary solution we add an option to deactivate autocomplete while keeping the address check active.

This option can be used by host system integrations.