Add an extandable advanced logic to field subscriber invalidation
Currently we we use the `hasActiveSubscriber` internally to check, if the subscribed DOM element exists and is not disabled. Because we don't want to
attempt to write or in generall process a form element, that is not usable. There is new problem case here: subdivision.

In some systems subdivision is entered through a select element, that is populated dynamically after country selection. The element itself is present and sometimes just hidden and without options, but our `hasActiveSubscriber` logik still recognizes it as valid form element and uses it for some proccesses such as addressCheck. And that creates an endless loop, where our solutions insists on correction, that cannot be applied.

In order to solve it two things are made:

1. The logic that checks subscriber validity should check options in a select. If they not have values or have values not present in mapping table for subdivisions, the field should be ignored

2. Allow extension of that logic to host system implementation, because they might need to add unique strategy