Check if the form and essential fields still exist in new address check
While refactoring the address check process some part got accidentally removed. They are responsible
for stopping the address check (or invalidating the results) if the receiver form or the essential fields
are not capable of receving.
Usually its the case in shopware 6, when the address in modal is closed before the results of address check are returned.

Ref: JS-119